requirements.txt
fastapi
uvicorn
python-multipart
python-docx
pymupdf
jinja2
spacy